Refactor: no ofrak_component dependencies (#21)
* modify instruction attributes in modifier so it can replace uimage modifier in tests

* move and update tests so ofrak_core does no depend on ofrak_components

* move pytest_plugins list to top-level conftest.py
