+++ ofrak_core/test_ofrak/components/conftest.py
@@ -1,7 +1,5 @@
 import pytest
 
-pytest_plugins = ["pytest_ofrak.fixtures", "pytest_ofrak.elf.fixtures"]
-
 
 @pytest.fixture
 def test_id():
