+++ ofrak_core/test_ofrak/conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = ["pytest_ofrak.fixtures"]
+pytest_plugins = ["pytest_ofrak.fixtures", "pytest_ofrak.elf.fixtures"]
